+++ branch-master/jslint.js
@@ -88,10 +88,12 @@
 /*jslint node*/
 
 /*property
-    unordered,
+    bind, cli, cwd, directive_quiet, endsWith, jslint, resolve, sep,
+    source_line,
+    unclosed_disable, unopened_enable, unordered,
     JSLINT_CLI, a, all, and, argv, arity, assign, b, bad_assignment_a,
     bad_directive_a, bad_get, bad_module_name_a, bad_option_a, bad_property_a,
-    bad_set, bitwise, block, body, browser, c, calls, catch, cli_mode, closer,
+    bad_set, bitwise, block, body, browser, c, calls, catch, closer,
     closure, code, column, concat, console_error, constant, context, convert,
     couch, create, d, dead, debug, default, devel, directive, directives,
     disrupt, dot, duplicate_a, early_stop, edition, ellipsis, else, empty_block,
@@ -102,7 +104,7 @@
     expected_space_a_b, expected_statements_a, expected_string_a,
     expected_type_string_a, exports, expression, extra, file, finally, flag,
     for, forEach, formatted_message, free, freeze, freeze_exports, from, froms,
-    fud, fudge, function_in_loop, functions, g, getset, global, has_await, i,
+    fud, function_in_loop, functions, g, getset, global, has_await, i,
     id, identifier, import, inc, indexOf, infix_in, init, initial, isArray,
     isNaN, is_async, join, json, keys, label, label_a, lbp, led, length, level,
     line, line_offset, lines, live, long, loop, m, map, margin, match, message,
@@ -121,7 +123,7 @@
     unexpected_directive_a, unexpected_expression_a, unexpected_label_a,
     unexpected_parens, unexpected_space_a_b, unexpected_statement_a,
     unexpected_trailing_space, unexpected_typeof_a, uninitialized_a,
-    unordered_param_a, unordered_property_a, unreachable_a,
+    unordered_a_b, unreachable_a,
     unregistered_property_a, unused_a, use_double, use_open, use_spaces, used,
     value, var_loop, var_switch, variable, versions, warning, warnings,
     weird_condition_a, weird_expression_a, weird_loop, weird_relation_a, white,
@@ -129,11 +131,11 @@
     wrapped, writable, y
 */
 
-const edition = "v2021.5.30";
+const edition = "v2021.6.3";
 
 function assert_or_throw(passed, message) {
 
-// this function will throw <message> if <passed> is falsy
+// This function will throw <message> if <passed> is falsy.
 
     if (passed) {
         return passed;
@@ -141,7 +143,8 @@ function assert_or_throw(passed, message) {
     throw new Error(`This was caused by a bug in JSLint.
 Please open an issue with this stack-trace (and possible example-code) at
 https://github.com/jslint-org/jslint/issues.
-edition = "${edition}";` + "\n" + message);
+edition = "${edition}";
+${message}`);
 }
 
 function empty() {
@@ -191,7 +194,6 @@ const allowed_option = {
     ],
     eval: true,
     for: true,
-    fudge: true,
     getset: true,
     long: true,
     node: [
@@ -328,6 +330,10 @@ const bundle = {
     too_long: "Line is longer than 80 characters.",
     too_many_digits: "Too many digits.",
     unclosed_comment: "Unclosed comment.",
+    unclosed_disable: (
+        "Directive '/*jslint-disable*/' was not closed "
+        + "with '/*jslint-enable*/'."
+    ),
     unclosed_mega: "Unclosed mega literal.",
     unclosed_string: "Unclosed string.",
     undeclared_a: "Undeclared '{a}'.",
@@ -352,8 +358,11 @@ const bundle = {
         "Unexpected 'typeof'. Use '===' to compare directly with {a}."
     ),
     uninitialized_a: "Uninitialized '{a}'.",
-    unordered_param_a: "Parameter '{a}' not listed in alphabetical order.",
-    unordered_property_a: "Property '{a}' not listed in alphabetical order.",
+    unopened_enable: (
+        "Directive '/*jslint-enable*/' was not opened "
+        + "with '/*jslint-disable*/'."
+    ),
+    unordered_a_b: "{a} '{b}' not listed in alphabetical order.",
     unreachable_a: "Unreachable '{a}'.",
     unregistered_property_a: "Unregistered property name '{a}'.",
     unused_a: "Unused '{a}'.",
@@ -388,11 +397,6 @@ function tag_regexp(strings) {
 
 // supplant {variables}
 const rx_supplant = /\{([^{}]*)\}/g;
-// carriage return, carriage return linefeed, or linefeed
-const rx_crlf = tag_regexp `
-      \n
-    | \r \n?
-`;
 // identifier
 const rx_identifier = tag_regexp ` ^(
     [ a-z A-Z _ $ ]
@@ -487,16 +491,16 @@ function is_letter(string) {
     );
 }
 
+const fudge = 1;        // Fudge starting line and starting column to 1.
 let anon;               // The guessed name for anonymous functions.
-let blockage;           // The current block.
 let block_stack;        // The stack of blocks.
+let blockage;           // The current block.
 let declared_globals;   // The object containing the global declarations.
-let directives;         // The directive comments.
 let directive_mode;     // true if directives are still allowed.
+let directives;         // The directive comments.
 let early_stop;         // true if JSLint cannot finish.
 let exports;            // The exported names and values.
 let froms;              // The array collecting all import-from strings.
-let fudge;              // true if the natural numbers start with 1.
 let functionage;        // The current function.
 let functions;          // The array containing all of the functions.
 let global;             // The global object; the outermost context.
@@ -510,10 +514,10 @@ let property;           // The object containing the tallied property names.
 let shebang;            // true if a #! was seen on the first line.
 let stack;              // The stack of functions.
 let syntax;             // The object containing the parser.
+let tenure;             // The predefined property registry.
 let token;              // The current token being examined in the parse.
 let token_nr;           // The number of the next token.
 let tokens;             // The array of tokens.
-let tenure;             // The predefined property registry.
 let tree;               // The abstract parse tree.
 let var_mode;           // "var" if using var; "let" if using let.
 let warnings;           // The array collecting all generated warnings.
@@ -539,24 +543,20 @@ function warn_at(code, line, column, a, b, c, d) {
 // Report an error at some line and column of the program. The warning object
 // resembles an exception.
 
-    const warning = {         // ~~
+    const warning = Object.assign({
+        a,
+        b,
+        c,
         code,
-        column,
+
+// Fudge column numbers in warning message.
+
+        column: column || fudge,
+        d,
         line,
-        name: "JSLintError"
-    };
-    if (a !== undefined) {
-        warning.a = a;
-    }
-    if (b !== undefined) {
-        warning.b = b;
-    }
-    if (c !== undefined) {
-        warning.c = c;
-    }
-    if (d !== undefined) {
-        warning.d = d;
-    }
+        name: "JSLintError",
+        source_line: ""
+    }, lines[line]);
     warning.message = bundle[code].replace(rx_supplant, function (
         ignore,
         filling
@@ -578,6 +578,12 @@ function warn_at(code, line, column, a, b, c, d) {
 
     if (option.debug) {
         warning.stack_trace = new Error().stack;
+    }
+    if (warning.directive_quiet) {
+
+// cause: "0 //jslint-quiet"
+
+        return warning;
     }
     warnings.push(warning);
     return warning;
@@ -603,7 +609,7 @@ function warn(code, the_token, a, b, c, d) {
         the_token.warning = warn_at(
             code,
             the_token.line,
-            the_token.from,
+            (the_token.from || 0) + fudge,
             a || artifact(the_token),
             b,
             c,
@@ -637,33 +643,37 @@ function tokenize(source) {
 // automatic semicolon insertion and nested megastring literals, which allows
 // full tokenization to precede parsing.
 
-// If the source is not an array, then it is split into lines at the
-// carriage return/linefeed.
-
-    lines = (
-        Array.isArray(source)
-        ? source
-        : source.split(rx_crlf)
-    );
-    tokens = [];
-
     let char;                   // a popular character
     let column = 0;             // the column number of the next character
+    let disable_line;           // the starting line of /*jslint-disable*/
     let first;                  // the first token
     let from;                   // the starting column number of the token
-    let line = -1;              // the line number of the next character
+    let line = 0;               // the line number of the next character
+    let mega_from;              // the starting column of megastring
+    let mega_line;              // the starting line of megastring
     let nr = 0;                 // the next token number
     let previous = global;      // the previous token including comments
     let prior = global;         // the previous token excluding comments
-    let mega_from;              // the starting column of megastring
-    let mega_line;              // the starting line of megastring
     let regexp_seen;            // regular expression literal seen on this line
     let snippet = "";           // a piece of string
     let source_line = "";       // the remaining line source string
     let whole_line = "";        // the whole line source string
 
-    if (lines[0].startsWith("#!")) {
-        line = 0;
+// Split source into lines at the carriage return/linefeed.
+
+    lines = String("\n" + source).split(
+        /\n|\r\n?/
+    ).map(function (source_line) {
+        return {
+            source_line
+        };
+    });
+    tokens = [];
+
+// Scan first line for "#!" and ignore it.
+
+    if (lines[fudge].source_line.startsWith("#!")) {
+        line += 1;
         shebang = true;
     }
 
@@ -677,6 +687,7 @@ function tokenize(source) {
         if (
             !option.long
             && whole_line.length > 80
+            && disable_line === undefined
             && !json_mode
             && first
             && !regexp_seen
@@ -689,26 +700,59 @@ function tokenize(source) {
         column = 0;
         line += 1;
         regexp_seen = false;
-        source_line = lines[line];
-        whole_line = source_line || "";
-        if (source_line !== undefined) {
-            at = source_line.search(rx_tab);
-            if (at >= 0) {
-                if (!option.white) {
+        source_line = undefined;
+        whole_line = "";
+        if (lines[line] === undefined) {
+            return source_line;
+        }
+        source_line = lines[line].source_line;
+        whole_line = source_line;
+
+// Scan each line for following ignore-directives:
+// "/*jslint-disable*/"
+// "/*jslint-enable*/"
+// "//jslint-quiet"
+
+        if (source_line === "/*jslint-disable*/") {
+
+// cause: "/*jslint-disable*/"
+
+            disable_line = line;
+        } else if (source_line === "/*jslint-enable*/") {
+            if (disable_line === undefined) {
+
+// cause: "/*jslint-enable*/"
+
+                stop_at("unopened_enable", line);
+            }
+            disable_line = undefined;
+        } else if (source_line.endsWith(" //jslint-quiet")) {
+
+// cause: "0 //jslint-quiet"
+
+            lines[line].directive_quiet = true;
+        }
+        if (disable_line !== undefined) {
+
+// cause: "/*jslint-disable*/\n0"
+
+            source_line = "";
+        }
+        at = source_line.search(rx_tab);
+        if (at >= 0) {
+            if (!option.white) {
 
 // cause: "\t"
 
-                    warn_at("use_spaces", line, at + 1);
-                }
-                source_line = source_line.replace(rx_tab, " ");
+                warn_at("use_spaces", line, at);
             }
-            if (!option.white && source_line.slice(-1) === " ") {
-                warn_at(
-                    "unexpected_trailing_space",
-                    line,
-                    source_line.length - 1
-                );
-            }
+            source_line = source_line.replace(rx_tab, " ");
+        }
+        if (!option.white && source_line.endsWith(" ")) {
+
+// cause: " "
+
+            warn_at("unexpected_trailing_space", line, source_line.length - 1);
         }
         return source_line;
     }
@@ -743,7 +787,7 @@ function tokenize(source) {
 
 // cause: "aa=/aa{/"
 
-                : stop_at("expected_a_b", line, column - 1, match, char)
+                : stop_at("expected_a_b", line, column, match, char)
             );
         }
         if (source_line) {
@@ -803,13 +847,13 @@ function tokenize(source) {
 
 // cause: "[\"\\u{12345}\"]"
 
-                    warn_at("unexpected_a", line, column - 1, char);
+                    warn_at("unexpected_a", line, column, char);
                 }
                 if (some_digits(rx_hexs) > 5) {
 
 // cause: "\"\\u{123456}\""
 
-                    warn_at("too_many_digits", line, column - 1);
+                    warn_at("too_many_digits", line, column);
                 }
                 if (char !== "}") {
 
@@ -824,7 +868,7 @@ function tokenize(source) {
 
 // cause: "\"\\u0\""
 
-                warn_at("expected_four_digits", line, column - 1);
+                warn_at("expected_four_digits", line, column);
             }
             return;
         }
@@ -832,9 +876,9 @@ function tokenize(source) {
             return next_char();
         }
 
-// cause: "\"\\a\""
+// cause: "\"\\0\""
 
-        warn_at("unexpected_a_before_b", line, column - 2, "\\", char);
+        warn_at("unexpected_a_before_b", line, column, "\\", char);
     }
 
     function make(id, value, identifier) {
@@ -1083,12 +1127,7 @@ function tokenize(source) {
 
 // cause: "aa=/[0-]/"
 
-                        return stop_at(
-                            "unexpected_a",
-                            line,
-                            column - 1,
-                            "-"
-                        );
+                        return stop_at("unexpected_a", line, column - 1, "-");
                     }
                 }
                 return ranges();
@@ -1109,13 +1148,7 @@ function tokenize(source) {
 
 // cause: "aa=/[/"
 
-                    warn_at(
-                        "expected_a_before_b",
-                        line,
-                        column,
-                        "\\",
-                        char
-                    );
+                    warn_at("expected_a_before_b", line, column, "\\", char);
                     next_char();
                     return classy();
                 }
@@ -1186,13 +1219,7 @@ function tokenize(source) {
                 case "*":
                 case "}":
                 case "{":
-                    warn_at(
-                        "expected_a_before_b",
-                        line,
-                        column - 1,
-                        "\\",
-                        char
-                    );
+                    warn_at("expected_a_before_b", line, column, "\\", char);
                     next_char();
                     break;
                 case "`":
@@ -1200,7 +1227,7 @@ function tokenize(source) {
 
 // cause: "`${/`/}`"
 
-                        warn_at("unexpected_a", line, column - 1, "`");
+                        warn_at("unexpected_a", line, column, "`");
                     }
                     next_char();
                     break;
@@ -1208,7 +1235,7 @@ function tokenize(source) {
 
 // cause: "aa=/ /"
 
-                    warn_at("expected_a_b", line, column - 1, "\\s", " ");
+                    warn_at("expected_a_b", line, column, "\\s", " ");
                     next_char();
                     break;
                 case "$":
@@ -1375,7 +1402,7 @@ function tokenize(source) {
             return stop_at(
                 "unexpected_a_after_b",
                 line,
-                column - 1,
+                column,
                 snippet.slice(-1),
                 snippet.slice(0, -1)
             );
@@ -1418,7 +1445,15 @@ function tokenize(source) {
             if (source_line === undefined) {
                 return (
                     mega_mode
+
+// cause: "`${//}`"
+
                     ? stop_at("unclosed_mega", mega_line, mega_from)
+                    : disable_line !== undefined
+
+// cause: "/*jslint-disable*/"
+
+                    ? stop_at("unclosed_disable", disable_line)
                     : make("(end)")
                 );
             }
@@ -1839,10 +1874,7 @@ function advance(id, match) {
                 next_token,
                 id,
                 artifact(match),
-
-// Return the fudged line number of an artifact.
-
-                match.line + fudge,
+                match.line,
                 artifact(next_token)
             )
         );
@@ -2002,7 +2034,7 @@ function enroll(name, role, readonly) {
                 "redefinition_a_b",
                 name,
                 name.id,
-                earlier.line + fudge
+                earlier.line
             );
 
 // Has the name been enrolled in an outer context?
@@ -2039,7 +2071,7 @@ function enroll(name, role, readonly) {
                             "redefinition_a_b",
                             name,
                             name.id,
-                            earlier.line + fudge
+                            earlier.line
                         );
                     }
                 }
@@ -3144,7 +3176,12 @@ function parameter_list() {
 
 // cause: "function aa({bb,aa}){}"
 
-                        warn("unordered_param_a", subparam);
+                        warn(
+                            "unordered_a_b",
+                            subparam,
+                            "Parameter",
+                            artifact(subparam)
+                        );
                     }
                     advance();
                     signature.push(subparam.id);
@@ -3573,7 +3610,7 @@ prefix("{", function () {
 
 // cause: "aa={bb,aa}"
 
-                warn("unordered_property_a", name);
+                warn("unordered_a_b", name, "Property", artifact(name));
             }
             if (
                 (name.id === "get" || name.id === "set")
@@ -3789,7 +3826,7 @@ function do_var() {
 
 // cause: "let{bb,aa}"
 
-                    warn("unordered_param_a", name);
+                    warn("unordered_a_b", name, "Parameter", artifact(name));
                 }
                 advance();
                 if (next_token.id === ":") {
@@ -4399,9 +4436,10 @@ stmt("throw", function () {
     return the_throw;
 });
 stmt("try", function () {
+    const the_try = token;
+    let ignored;
     let the_catch;
     let the_disrupt;
-    const the_try = token;
     if (functionage.try > 0) {
 
 // cause: "try{try{}catch(){}}catch(){}"
@@ -4412,10 +4450,16 @@ stmt("try", function () {
     the_try.block = block();
     the_disrupt = the_try.block.disrupt;
     if (next_token.id === "catch") {
-        let ignored = "ignore";
-        the_catch = next_token;
-        the_try.catch = the_catch;
         advance("catch");
+        the_catch = next_token;
+        ignored = "ignore";
+        the_try.catch = the_catch;
+
+// Create new function-scope for catch-parameter.
+
+        stack.push(functionage);
+        functionage = the_catch;
+        functionage.context = empty();
         if (next_token.id === "(") {
             advance("(");
             if (!next_token.identifier) {
@@ -4436,6 +4480,10 @@ stmt("try", function () {
         if (the_catch.block.disrupt !== true) {
             the_disrupt = false;
         }
+
+// Restore previous function-scope after catch-block.
+
+        functionage = stack.pop();
     } else {
 
 // cause: "try{}finally{break;}"
@@ -4677,6 +4725,7 @@ function lookup(thing) {
 
 // cause: "aa"
 // cause: "class aa{}"
+// cause: "let aa=0;try{aa();}catch(bb){bb();}bb();"
 
                     warn("undeclared_a", thing);
                     return;
@@ -4959,6 +5008,15 @@ preaction("statement", "for", function (thing) {
     walk_statement(thing.initial);
 });
 preaction("statement", "function", preaction_function);
+preaction("statement", "try", function (thing) {
+    if (thing.catch !== undefined) {
+
+// Create new function-scope for catch-parameter.
+
+        stack.push(functionage);
+        functionage = thing.catch;
+    }
+});
 preaction("unary", "~", bitwise_check);
 preaction("unary", "function", preaction_function);
 preaction("variable", function (thing) {
@@ -5360,6 +5418,10 @@ postaction("statement", "try", function (thing) {
             the_variable.init = true;
         }
         walk_statement(thing.catch.block);
+
+// Restore previous function-scope after catch-block.
+
+        functionage = stack.pop();
     }
 });
 postaction("statement", "var", action_var);
@@ -5573,10 +5635,10 @@ function whitage() {
             "expected_a_at_b_c",
             right,
             artifact(right),
-            fudge + at,
 
-// Return the fudged column number of an artifact.
+// Fudge column numbers in warning message.
 
+            at + fudge,
             right.from + fudge
         );
     }
@@ -5597,33 +5659,6 @@ function whitage() {
                 || left.thru !== right.from
             )
         ) {
-
-// cause:
-//  (
-// function aa (
-// bb
-// ,
-// [
-// cc,dd
-// ]
-// ,
-// {
-// ee,ff=( 0 )
-// }
-// ,
-// ... zz
-// )
-// {
-// return {
-// aa,bb
-// }
-// ;
-// }
-// (
-// )
-//  )
-//  ;
-
             warn(
                 "unexpected_space_a_b",
                 right,
@@ -6081,11 +6116,6 @@ function jslint(
         early_stop = true;
         exports = empty();
         froms = [];
-        fudge = (
-            option.fudge
-            ? 1
-            : 0
-        );
         functions = [];
         global = {
             body: true,
@@ -6094,7 +6124,7 @@ function jslint(
             from: 0,
             id: "(global)",
             level: 0,
-            line: 0,
+            line: fudge,
             live: [],
             loop: 0,
             switch: 0,
@@ -6124,7 +6154,13 @@ function jslint(
                 }
             }
         });
+
+// 1. Tokenize source into ast.
+
         tokenize(source);
+
+// 2. Walk ast.
+
         advance();
         if (json_mode) {
             tree = json_value();
@@ -6153,6 +6189,9 @@ function jslint(
             advance("(end)");
             functionage = global;
             walk_statement(tree);
+
+// 3. Re-walk ast validating whitespace.
+
             if (warnings.length === 0) {
                 uninitialized_and_unused();
                 if (!option.white) {
@@ -6178,9 +6217,10 @@ function jslint(
         e.early_stop = true;
         e.message = "[JSLint was unable to finish]\n" + e.message;
         if (e.name !== "JSLintError") {
-            e.column = 0;
-            e.line = 0;
+            e.column = fudge;
+            e.line = fudge;
             e.stack_trace = e.stack;
+            e.source_line = "";
             warnings.push(e);
         }
     }
@@ -6196,20 +6236,19 @@ function jslint(
 // update each warning with a formatted_message ready for use by cli
 
     }).map(function ({
-        column = 0,
-        line = 0,
-        message = "",
+        column,
+        line,
+        message,
+        source_line,
         stack_trace = ""
     }, ii, list) {
-        column += 1;
-        line += 1;
         list[ii].formatted_message = String(
             String(ii + 1).padStart(3, " ")
             + " \u001b[31m" + message + "\u001b[39m"
             + " \u001b[90m\/\/ line " + line + ", column " + column
             + "\u001b[39m\n"
-            + ("    " + String(lines[line - 1] || "").trim()).slice(0, 72)
-            + "\n" + stack_trace
+            + ("    " + source_line.trim()).slice(0, 72) + "\n"
+            + stack_trace
         ).trim();
     });
     return {
@@ -6228,7 +6267,7 @@ function jslint(
         property,
         shebang: (
             shebang
-            ? lines[0]
+            ? lines[fudge].source_line
             : undefined
         ),
         stop: early_stop,
@@ -6244,18 +6283,23 @@ async function cli({
     option,
     source
 }) {
-/*
- * this function will run jslint from nodejs-cli
- */
+
+// This function will run jslint from nodejs-cli.
+
     const fs = await import("fs");
-    let exitCode;
+    const path = await import("path");
+    let data;
+    let exit_code;
+
     function string_line_count(code) {
-    /*
-     * this function will count number of newlines in <code>
-     */
+
+// This function will count number of newlines in <code>.
+
         let cnt;
         let ii;
-        // https://jsperf.com/regexp-counting-2/8
+
+// https://jsperf.com/regexp-counting-2/8
+
         cnt = 0;
         ii = 0;
         while (true) {
@@ -6267,6 +6311,7 @@ async function cli({
         }
         return cnt;
     }
+
     function jslint_from_file({
         code,
         file,
@@ -6278,7 +6323,9 @@ async function cli({
             /\.\w+?$|$/m
         ).exec(file)[0]) {
         case ".html":
-            // recurse
+
+// Recursively jslint embedded "<script>\n...\n</script>".
+
             code.replace((
                 /^<script>\n([\S\s]*?\n)<\/script>$/gm
             ), function (ignore, match1, ii) {
@@ -6294,7 +6341,9 @@ async function cli({
             });
             return;
         case ".md":
-            // recurse
+
+// Recursively jslint embedded "```javascript\n...\n```".
+
             code.replace((
                 /^```javascript\n([\S\s]*?\n)```$/gm
             ), function (ignore, match1, ii) {
@@ -6308,7 +6357,9 @@ async function cli({
             });
             return;
         case ".sh":
-            // recurse
+
+// Recursively jslint embedded "node -e '\n...\n'".
+
             code.replace((
                 /\bnode\u0020.*?-e\u0020'\n([\S\s]*?\n)'/gm
             ), function (ignore, match1, ii) {
@@ -6329,10 +6380,11 @@ async function cli({
                 option
             ).warnings;
         }
-        // print only first 10 warnings
+
+// Print only first 10 warnings to stderr.
+
         if (warnings.length > 0) {
-            exitCode = 1;
-            // print first 10 warnings to stderr
+            exit_code = 1;
             console_error(
                 "\u001b[1mjslint " + file + "\u001b[22m\n"
                 + warnings.slice(0, 10).map(function ({
@@ -6343,7 +6395,18 @@ async function cli({
             );
         }
     }
+
     console_error = console_error || console.error;
+
+// Normalize file relative to process.cwd().
+
+    file = path.resolve(file) + path.sep;
+    if (file.startsWith(process.cwd() + path.sep)) {
+        file = file.replace(process.cwd() + path.sep, "").slice(0, -1) || ".";
+    }
+    file = file.replace((
+        /\\/g
+    ), "/");
     if (source) {
         jslint_from_file({
             code: source,
@@ -6352,14 +6415,20 @@ async function cli({
         });
         return;
     }
-    if (file === ".") {
-        file = await fs.promises.readdir(".");
-        await Promise.all(file.map(async function (file) {
+
+// cli - jslint directory.
+
+    try {
+        data = await fs.promises.readdir(file, "utf8");
+    } catch (ignore) {}
+    if (data) {
+        await Promise.all(data.map(async function (file2) {
             let code;
-            let timeStart = Date.now();
+            let time_start = Date.now();
+            file2 = file + "/" + file2;
             switch ((
                 /\.\w+?$|$/m
-            ).exec(file)[0]) {
+            ).exec(file2)[0]) {
             case ".html":
             case ".js":
             case ".json":
@@ -6371,63 +6440,71 @@ async function cli({
                 return;
             }
             try {
-                code = await fs.promises.readFile(file, "utf8");
+                code = await fs.promises.readFile(file2, "utf8");
             } catch (ignore) {
                 return;
             }
             if (!(
                 !(
                     /\b(?:lock|min|raw|rollup)\b/
-                ).test(file) && code && code.length < 1048576
+                ).test(file2) && code && code.length < 1048576
             )) {
                 return;
             }
             jslint_from_file({
                 code,
-                file,
+                file: file2,
                 option
             });
             console_error(
-                "jslint - " + (Date.now() - timeStart) + "ms - " + file
+                "jslint - " + (Date.now() - time_start) + "ms - " + file2
             );
         }));
-    } else {
-        jslint_from_file({
-            code: await fs.promises.readFile(file, "utf8"),
-            file,
-            option
-        });
+        return exit_code;
     }
-    return exitCode;
+
+// cli - jslint file.
+
+    try {
+        data = await fs.promises.readFile(file, "utf8");
+    } catch (err) {
+        console_error(err);
+        return 1;
+    }
+    jslint_from_file({
+        code: data,
+        file,
+        option
+    });
+    return exit_code;
 }
-export default Object.freeze(function (
-    source = "",
-    option_object = empty(),
-    global_array = []
-) {
-    if (option_object.cli_mode) {
-        return cli(Object.assign({
-            source
-        }, option_object));
-    }
-    return jslint(source, option_object, global_array);
-});
+jslint.cli = Object.freeze(cli);
+jslint.jslint = Object.freeze(jslint.bind(undefined));
+export default Object.freeze(jslint);
+
 // feature-detect nodejs-cli
+
 if (
     typeof process === "object"
-    // uncomment when nodejs v12 is no longer used in ci
-    // && typeof process?.versions?.node === "string"
+
+//  uncomment when nodejs v12 is no longer used in ci
+//  && typeof process?.versions?.node === "string"
+
     && process && process.versions
     && typeof process.versions.node === "string"
     && (
-        (/[\/|\\]jslint.m?js$/m).test(process.argv[1])
+        (
+            /[\/|\\]jslint\.m?js$/m
+        ).test(process.argv[1])
         || process.env.JSLINT_CLI === "1"
     )
 ) {
-    // run cli
+
+// run cli
+
     cli({
         file: process.argv[2]
-    }).then(function (exitCode) {
-        process.exit(exitCode);
+    }).then(function (exit_code) {
+        process.exit(exit_code);
     });
 }